Fix L2 packet receive
Below issues are seen with packet sockets:

* Despite using UDP sockets the networking stack gives the packet along
with L2 (ethernet) header, but as per [1] it should not work like
that. As the wpa_supplicant expects only the payload, so, add a check
and discard the L2 header.

  - As a side effect sll_addr is also not being filled.

* Despite binding to a specific protocol, other messages are also
received, probably due to the default bind in the networking stack, so,
add a check to explicitly drop non-matching frames.

[1] - https://docs.zephyrproject.org/3.0.0/reference/kconfig/CONFIG_NET_SOCKETS_PACKET_DGRAM.html