[nrf toup] zephyr: Disable capping the SNR
Any SNR greater than 25 is considered good enough and is capped at that
and then the scan results are sorted based on other factors. This works
well in networks which are close, but in case if networks are fanned
out, then we might end up choosing a bad network even though its SNR is
better.

Remove this cap and use the full SNR for comparison as well.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>